[DeltaCounter] renaming everything to **DeltaCounter and making class comments english. Full localization should be done in separate task I think.
