+++ src/Qoollo.PerformanceCounters/Generic/Categories/SingleInstanceCategory.cs
@@ -3,93 +3,90 @@
 namespace Qoollo.PerformanceCounters
 {
     /// <summary>
-    /// Создание категории с одним инстансом
+    /// Counter category with a single instance
     /// </summary>
     public abstract class SingleInstanceCategory : Category
     {
         /// <summary>
-        /// Конструктор
+        /// Constructor
         /// </summary>
-        /// <param name="categoryName">Имя категории</param>
-        /// <param name="categoryDescription">Описание категории</param>
+        /// <param name="categoryName">Category name</param>
+        /// <param name="categoryDescription">Category description</param>
         protected SingleInstanceCategory(string categoryName, string categoryDescription)
             : base(categoryName, categoryDescription, CategoryTypes.SingleInstance)
         {
         }
 
-
-        #region Работа со счетчиками
-
         /// <summary>
-        /// Создать счётчик определённого типа
+        /// Create counter
         /// </summary>
-        /// <param name="type">Тип счётчика</param>
-        /// <param name="counterName">Имя счётчика</param>
-        /// <param name="counterDescription">Описание счётчика</param>
-        /// <returns>Счётчик</returns>
+        /// <param name="type">Counter type</param>
+        /// <param name="counterName">Existing counter name</param>
+        /// <param name="counterDescription">Counter description</param>
+        /// <returns>Counter</returns>
         public abstract Counter CreateCounter(CounterTypes type, string counterName, string counterDescription);
 
         /// <summary>
-        /// Создать счётчик определённого типа (NumberOfItemsCounter)
+        /// Create counter (NumberOfItemsCounter)
         /// </summary>
-        /// <param name="counterName">Имя счетчика</param>
-        /// <param name="counterDescription">Описание счетчика</param>
-        /// <returns>Счётчик</returns>
+        /// <param name="counterName">Existing counter name</param>
+        /// <param name="counterDescription">Counter description</param>
+        /// <returns>Counter</returns>
         public virtual NumberOfItemsCounter CreateNumberOfItemsCounter(string counterName, string counterDescription)
         {
             return (NumberOfItemsCounter)CreateCounter(CounterTypes.NumberOfItems, counterName, counterDescription);
         }
 
         /// <summary>
-        /// Создать счётчик определённого типа (NumberOfItemsCounter)
+        /// Create counter (DeltaCounter)
         /// </summary>
-        /// <param name="counterName">Имя счетчика</param>
-        /// <param name="counterDescription">Описание счетчика</param>
-        /// <returns>Счётчик</returns>
-        public virtual DeltaCountCounter CreateDeltaCountCounter(string counterName, string counterDescription)
+        /// <param name="counterName">Existing counter name</param>
+        /// <param name="counterDescription">Counter description</param>
+        /// <returns>Counter</returns>
+        public virtual DeltaCounter CreateDeltaCounter(string counterName, string counterDescription)
         {
-            return (DeltaCountCounter)CreateCounter(CounterTypes.DeltaCount, counterName, counterDescription);
+            return (DeltaCounter)CreateCounter(CounterTypes.Delta, counterName, counterDescription);
         }
 
         /// <summary>
-        /// Создать счётчик определённого типа (AverageTimeCounter)
+        /// Create counter (AverageTimeCounter)
         /// </summary>
-        /// <param name="counterName">Имя счетчика</param>
-        /// <param name="counterDescription">Описание счетчика</param>
-        /// <returns>Счётчик</returns>
+        /// <param name="counterName">Existing counter name</param>
+        /// <param name="counterDescription">Counter description</param>
+        /// <returns>Counter</returns>
         public virtual AverageTimeCounter CreateAverageTimeCounter(string counterName, string counterDescription)
         {
             return (AverageTimeCounter)CreateCounter(CounterTypes.AverageTime, counterName, counterDescription);
         }
 
         /// <summary>
-        /// Создать счётчик определённого типа (AverageCountCounter)
+        /// Create counter (AverageCountCounter)
         /// </summary>
-        /// <param name="counterName">Имя счетчика</param>
-        /// <param name="counterDescription">Описание счетчика</param>
-        /// <returns>Счётчик</returns>
+        /// <param name="counterName">Existing counter name</param>
+        /// <param name="counterDescription">Counter description</param>
+        /// <returns>Counter</returns>
         public virtual AverageCountCounter CreateAverageCountCounter(string counterName, string counterDescription)
         {
             return (AverageCountCounter)CreateCounter(CounterTypes.AverageCount, counterName, counterDescription);
         }
 
         /// <summary>
-        /// Создать счётчик определённого типа (OperationsPerSecondCounter)
+        /// Create counter (OperationsPerSecondCounter)
         /// </summary>
-        /// <param name="counterName">Имя счетчика</param>
-        /// <param name="counterDescription">Описание счетчика</param>
-        /// <returns>Счётчик</returns>
+        /// <param name="counterName">Existing counter name</param>
+        /// <param name="counterDescription">Counter description</param>
+        /// <returns>Counter</returns>
         public virtual OperationsPerSecondCounter CreateOperationsPerSecondCounter(string counterName, string counterDescription)
         {
             return (OperationsPerSecondCounter)CreateCounter(CounterTypes.OperationsPerSecond, counterName, counterDescription);
         }
 
         /// <summary>
-        /// Создать счётчик определённого типа (ElapsedTimeCounter)
+        /// Create counter (ElapsedTimeCounter)
         /// </summary>
-        /// <param name="counterName">Имя счетчика</param>
-        /// <param name="counterDescription">Описание счетчика</param>
-        /// <returns>Счётчик</returns>
+        /// <param name="counterName">Existing counter name</param>
+        /// <param name="counterDescription">Counter description</param>
+        /// <returns>Counter</returns>
         public virtual ElapsedTimeCounter CreateElapsedTimeCounter(string counterName, string counterDescription)
         {
             return (ElapsedTimeCounter)CreateCounter(CounterTypes.ElapsedTime, counterName, counterDescription);
@@ -97,11 +94,11 @@ namespace Qoollo.PerformanceCounters
 
 
         /// <summary>
-        /// Создать счётчик определённого типа (MomentTimeCounter)
+        /// Create counter (MomentTimeCounter)
         /// </summary>
-        /// <param name="counterName">Имя счетчика</param>
-        /// <param name="counterDescription">Описание счетчика</param>
-        /// <returns>Счётчик</returns>
+        /// <param name="counterName">Existing counter name</param>
+        /// <param name="counterDescription">Counter description</param>
+        /// <returns>Counter</returns>
         public virtual MomentTimeCounter CreateMomentTimeCounter(string counterName, string counterDescription)
         {
             return (MomentTimeCounter)CreateCounter(CounterTypes.MomentTime, counterName, counterDescription);
@@ -109,52 +106,52 @@ namespace Qoollo.PerformanceCounters
 
 
         /// <summary>
-        /// Есть ли счётчик с указанным именем
+        /// checks existence of counter with specific name
         /// </summary>
-        /// <param name="counterName">Имя счётчика</param>
-        /// <returns>Есть ли он</returns>
+        /// <param name="counterName">Existing counter name</param>
+        /// <returns>True if exists</returns>
         public abstract bool HasCounter(string counterName);
 
         /// <summary>
-        /// Получение счетчика определенного типа
+        /// Returns existing counter by name
         /// </summary>
-        /// <param name="counterName">Имя счетчика</param>
-        /// <returns>Счётчик</returns>
+        /// <param name="counterName">Existing counter name</param>
+        /// <returns>Counter</returns>
         public abstract Counter GetCounter(string counterName);
 
         /// <summary>
-        /// Получение счетчика определенного типа
+        /// Returns existing counter by name and type
         /// </summary>
-        /// <param name="counterName">Имя счетчика</param>
-        /// <param name="expectedCounterType">Тип счётчика</param>
-        /// <returns>Счётчик</returns>
+        /// <param name="counterName">Existing counter name</param>
+        /// <param name="expectedCounterType">Counter type</param>
+        /// <returns>Counter</returns>
         public abstract Counter GetCounter(string counterName, CounterTypes expectedCounterType);
 
         /// <summary>
-        /// Возвращает ранее созданный счетчик производительности (NumberOfItemsCounter)
+        /// Returns existing NumberOfItemsCounter by name
         /// </summary>
-        /// <param name="counterName">Имя счетчика</param>
-        /// <returns>Счётчик</returns>
+        /// <param name="counterName">Existing counter name</param>
+        /// <returns>Counter</returns>
         public virtual NumberOfItemsCounter GetNumberOfItemsCounter(string counterName)
         {
             return (NumberOfItemsCounter)GetCounter(counterName, CounterTypes.NumberOfItems);
         }
 
         /// <summary>
-        /// Возвращает ранее созданный счетчик производительности (NumberOfItemsCounter)
+        /// Returns existing DeltaCounter by name
         /// </summary>
-        /// <param name="counterName">Имя счетчика</param>
-        /// <returns>Счётчик</returns>
-        public virtual DeltaCountCounter GetDeltaCountCounter(string counterName)
+        /// <param name="counterName">Existing counter name</param>
+        /// <returns>Counter</returns>
+        public virtual DeltaCounter GetDeltaCounter(string counterName)
         {
-            return (DeltaCountCounter)GetCounter(counterName, CounterTypes.DeltaCount);
+            return (DeltaCounter)GetCounter(counterName, CounterTypes.Delta);
         }
 
         /// <summary>
-        /// Возвращает ранее созданный счетчик производительности (AverageTimeCounter)
+        /// Returns existing AverageTimeCounter by name
         /// </summary>
-        /// <param name="counterName">Имя счетчика</param>
-        /// <returns>Счётчик</returns>
+        /// <param name="counterName">Existing counter name</param>
+        /// <returns>Counter</returns>
         public virtual AverageTimeCounter GetAverageTimeCounter(string counterName)
         {
             return (AverageTimeCounter)GetCounter(counterName, CounterTypes.AverageTime);
@@ -162,20 +159,20 @@ namespace Qoollo.PerformanceCounters
 
 
         /// <summary>
-        /// Возвращает ранее созданный счетчик производительности (AverageCountCounter)
+        /// Returns existing AverageCountCounter by name
         /// </summary>
-        /// <param name="counterName">Имя счетчика</param>
-        /// <returns>Счётчик</returns>
+        /// <param name="counterName">Existing counter name</param>
+        /// <returns>Counter</returns>
         public virtual AverageCountCounter GetAverageCountCounter(string counterName)
         {
             return (AverageCountCounter)GetCounter(counterName, CounterTypes.AverageCount);
         }
 
         /// <summary>
-        /// Возвращает ранее созданный счетчик производительности (OperationsPerSecondCounter)
+        /// Returns existing OperationsPerSecondCounter by name
         /// </summary>
-        /// <param name="counterName">Имя счетчика</param>
-        /// <returns>Счётчик</returns>
+        /// <param name="counterName">Existing counter name</param>
+        /// <returns>Counter</returns>
         public virtual OperationsPerSecondCounter GetOperationsPerSecondCounter(string counterName)
         {
             return (OperationsPerSecondCounter)GetCounter(counterName, CounterTypes.OperationsPerSecond);
@@ -183,27 +180,24 @@ namespace Qoollo.PerformanceCounters
 
 
         /// <summary>
-        /// Возвращает ранее созданный счетчик производительности (ElapsedTimeCounter)
+        /// Returns existing ElapsedTimeCounter by name
         /// </summary>
-        /// <param name="counterName">Имя счетчика</param>
-        /// <returns>Счётчик</returns>
+        /// <param name="counterName">Existing counter name</param>
+        /// <returns>Counter</returns>
         public virtual ElapsedTimeCounter GetElapsedTimeCounter(string counterName)
         {
             return (ElapsedTimeCounter)GetCounter(counterName, CounterTypes.ElapsedTime);
         }
 
         /// <summary>
-        /// Возвращает ранее созданный счетчик производительности (MomentTimeCounter)
+        /// Returns existing MomentTimeCounter by name
         /// </summary>
-        /// <param name="counterName">Имя счетчика</param>
-        /// <returns>Счётчик</returns>
+        /// <param name="counterName">Existing counter name</param>
+        /// <returns>Counter</returns>
         public virtual MomentTimeCounter GetMomentTimeCounter(string counterName)
         {
             return (MomentTimeCounter)GetCounter(counterName, CounterTypes.MomentTime);
         }
-
-
-        #endregion
 
     }
 }
