+++ src/Qoollo.PerformanceCounters/Generic/Counters/CounterTypeConverter.cs
@@ -24,8 +24,8 @@ namespace Qoollo.PerformanceCounters
             if (type == typeof(NumberOfItemsCounter))
                 return CounterTypes.NumberOfItems;
 
-            if (type == typeof(DeltaCountCounter))
-                return CounterTypes.DeltaCount;
+            if (type == typeof(DeltaCounter))
+                return CounterTypes.Delta;
 
             if (type == typeof(AverageTimeCounter))
                 return CounterTypes.AverageTime;
