+++ src/Qoollo.PerformanceCounters/InternalCounters/Categories/InternalSingleInstanceCategory.cs
@@ -69,8 +69,8 @@ namespace Qoollo.PerformanceCounters.InternalCounters.Categories
                 case CounterTypes.NumberOfItems:
                     res = new InternalNumberOfItemsCounter(counterName, counterDescription);
                     break;
-                case CounterTypes.DeltaCount:
-                    res = new InternalDeltaCountCounter(counterName, counterDescription);
+                case CounterTypes.Delta:
+                    res = new InternalDeltaCounter(counterName, counterDescription);
                     break;
                 case CounterTypes.AverageTime:
                     res = new InternalAverageTimeCounter(counterName, counterDescription);
